some logging cleanup rewrite service-token cache logic to be thread safe, and implement renewal attempts when the service token gets near the end of its lifetime.
This behavior is controlled by two #defines in mod_webauth.h,  one
define controls when the first retry is (90% of the token's lifetime),
and the other is how long to wait after a failed attempt (10 minutes)
before we try again. We keep the valid token around if the renewal failed
of course. These values are currently hardcoded, probably don't need
to be configs...