Add libtool support.  Build shared and static libwebauth by default.  Fix the Perl bindings to build correctly against a shared libwebauth, if it exists.
Add --with-krb5 and --with-crypto arguments to configure to specify the
roots of the Kerberos v5 and OpenSSL installations, if they're not in
locations found by default by the compiler.

Remove the check for the actual SSL library (as opposed to just the crypto
library) for right now; it will be added back in in a different way when
the module build is integrated into the build system.