add WebAuthKeyringAutoUdpate and WebAuthKeyringKeyLifetime to automatically manage the keyring file.
currently only partially-implemented in mod_webauth. We'll create
the keyring file if its not there, but its currently not updated,
this will be finished tomorrow.

also, we still might decide to remove the valid_till field on keyring
entries so they don't expire, though the key with the more current valid_from
date that is not-posted would always be the one used for encrypting new
data.