Integrate the build of the Apache modules into the rest of the build system using apxs to build them.  make install support isn't present yet, so for the time being apxs has to still be called directly to do installs.
Adds --with-ssl and --with-curl to specify the locations to OpenSSL and
cURL.  Uses the --with-apache option included in the last change.  Try to
have Apache's libtool link properly against the uninstalled libwebauth.la
libtool library (it's not clear that this is working right yet).

Removes --with-crypto in favor of just having --with-ssl that finds both
the SSL library and the crypto library.