Significant cleanup to match Perl coding standards, simplify, remove commented-out code, reformat, and add many comments.
Modified the logic of login so that we wouldn't always print the login
screen if the test cookie wasn't found, since the test cookie is now a
session cookie.  Instead, only warn about missing cookie support if the
user logged in with a username and password, since that's the best that
we can do right now.