General terminology change from spnego to remuser, with corresponding changes to all variables, functions, cookie names, and template variables.
Update the comment block in the templates that says what variables are
used and exclude that comment from the output HTML (no need to expose
it to the user).  Standardize the formatting of that block.

Add an error variable for the login.fcgi page that is set if any error
is present, to make templating the login screen easier.  Rename some
other template variables for general consistency, since we're breaking
old templates anyway.

Don't provide an option for Apache authentication if the user already
tried to log in and failed.  Otherwise, we may provide one even if the
user is being forced to log in, which would then not work and be
confusing.  Also, don't provide the option for Apache authentication
if $REMOTE_USER_REDIRECT isn't configured; people may still want to
use the old method of always trying Apache authentication first.

Lots of additional commenting and similar code cleanup fixes.