Improvements to WebLogin confirmation screen bypass
Setting $BYPASS_CONFIRM in the WebLogin configuration now also
suppresses the confirmation page after username/password login
provided that the browser supports HTTP/1.1 (and the web server tells
the WebLogin script that in the form Apache does).

WebLogin now sets and updates its cookies after successful
authentication even if the confirmation screen is bypassed.  This
primarily affects the update of the expiration time of the REMOTE_USER
cookie.