Additional WebLogin confirmation page detail for cred delegation
Based on the requested token type, always show the confirmation page
in WebLogin if the WAS requested a proxy token (thus allowing it to
get delegated credentials).  Add a new template variable for the
confirmation page that includes information about the credentials that
host is allowed to request.  Add a new WebLogin configuration variable
specifying the path to the token.acl file to gather that information.
Recognize "id" as a special $BYPASS_CONFIRM setting to show the
confirm page only if the WAS requested proxy tokens.