Massive renaming of the source for a more standard layout
Eliminate the src/* level of directory hierarchy.  Move libwebauth to
just lib.  Move bindings/perl/WebAuth to perl and merge it with the
WebKDC Perl modules.  Move src/webkdc to weblogin, since that's what it
truly is.

Eliminate the remnants of the mkmanifest way of creating distributions.
Remove from Git all the Automake, Autoconf, and Libtool supporting files,
which will be added by autogen instead.  Reuse the tools directory for
wa_keyring.

Rename doc to docs to match other projects.