Fix installation rules for data files
Stop installing the test suite for the time being since Automake makes
it extremely difficult to install a nested static directory structure.
This will be re-added when the test suite is rewritten to require fewer
separate directories.

List the WebLogin templates and images to install separately, since
directories cannot be listed as install sources.  Install only the PNG
images, not the XCF images, but list the latter in EXTRA_DIST.