Various updates to INSTALL and test procedures
We no longer have a target to install the test suite.  Instead, describe
how to copy the test suite to the appropriate location.  Update NEWS
accordingly.

Update the install instructions to remove some obsolete bits about pubsw,
change CVS to Git, and update required software versions.  Note that the
keyring and keytab can go wherever, as long as the Apache configuration is
updated.

Update the ssl FAQ pointer to the Apache 2.2 documentation.