More installation fixes
Move the WebKDC* modules into a separate tree during the installation
rather than installing them from /usr/lib/perl5 since that ends up
with only arch-independent files in /usr/lib/perl5 in the libwebkdc-perl
package.

Install the Apache modules with libtool in the hope that this will
eliminate the rpaths that they're ending up with.