WebLogin: Assume cookie support on POST of the login form
If the user submits the login form via POST without including the test
cookie, assume that the browser supports cookies and proceed.  We
won't present the initial login form without seeing the test cookie,
so something strange is happening.  Continuing and assuming everything
will work seems to be the best approach.