Bring source tree in line with tarball
Revert lots of changes caused by conversions between revision control
systems and the fact we're basing the source tree on what was in the
revision control system rather than the release.  This removes the
files and directory we didn't distribute and puts back all of the Id
strings matching what was in Subversion at the time.