perl/lib/WebKDC/Token.pm: Updated documentation for $ttl
The purpose of the $ttl field when parsing an encoded existing token was
specified in the C libraries that parse_token calls, but not the Perl
module.  Added a short note about usage in that case to the module, and
fixed the use of a non-existant fourth argument in the example.