Import C TAP Harness 1.3 and rewrite default tests
Update the C test framework to C TAP Harness 1.3 and import the basic
TAP library from there.  Rename all test programs to use the -t
extension instead of _test.  Convert the tests that are run by default
to using the TAP library instead of the set of macros included with
WebAuth.