Use strdup and check results instead of malloc/strcpy
In several places in the WebAuth library code, we were setting the
client principal after krb5_aname_to_localname with malloc plus
strcpy and not checking the results of malloc.  Switch to strdup
instead and check that we were able to allocate memory.  This removes
all remaining uses of strcpy.