Fixes for expiring password remctl access for testing
Added functionality to perl/lib/WebLogin.pm to take a different port and
server principal to run remctl with.  Added the configuration to
perl/lib/WebKDC/Config.pm and documented in docs/weblogin-config.  These
are required for testing, where we may run a fake remctl server with
non-standard port and principal, but could be useful in other cases.

Also fixed minor bug with password not being properly checked before
creating a CPT when CPT did not exist.