Rewrite the cURL probes to use better standard practices
The Autoconf probe for the cURL libraries now uses curl-config if
available.  The path to curl-config can be overridden by setting the
CURL_CONFIG variable on the configure command line or in the
environment.

Add macros to switch and restore flags for cURL so that we can do
library probes later if we need to.