Rationalize privgroup authrule behavior, fix documentation mismatch
mod_webauthldap's behavior did not match the documentation for the
directive WebAuthLdapAuthrule.  While the docs claimed that the
WEBAUTH_LDAPAUTHRULE environment variable was set to "group <privgroup>"
for privgroup matches, in fact the variable was set to just the privgroup
name.  A saner behavior, implemented here, is to set the variable to the
Require-like string "privgroup <privgroup>" to parallel the other cases,
"valid-user" and "user <user>".

This change is *not* backwards-compatible.  Developers using
WEBAUTH_LDAPAUTHRULE may have to update their web applications.

While making this change, consolidate authrule-setting code in the
privgroup validation function; this refactor allows webauthldap_docompare
to be reused for arbitrary privgroup-checking functionality.