Add WebAuthLdapPrivgroup to mod_webauthldap
The WebAuthLdapPrivgroup directive allows the remote user's membership
in arbitrary privgroups to be checked in LDAP, without requiring
that the user be a member of any group.  It is very similar to
WebAuthLdapAttribute, but it operates exclusively on the particular
attribute specified by WebAuthLdapAuthorizationAttribute, and each
group to be checked must be specified explicitly as an argument to the
directive.  The membership checks are performed via LDAP comparisons,
so the feature is useful for environments in which the web application
server's principal is only granted compare access to the authorization
attribute.