Improve error reporting in WebLogin when password change fails
Handle the case of an invalid username or password explicitly and
activate the template path that shows a dedicated error for that.
Produce a better error message when the change password token has
timed out.  Include the error message from the kpasswdd server if
password change fails, but adjust it by removing some boilerplate
that isn't useful for users.