Work around a password change bug with MIT Kerberos
Previous versions of WebLogin interpreted a "message stream modified"
error on password change as a failure of strength checking because
that error was incorrectly returned by MIT Kerberos for password
strength checking errors with a Heimdal KDC.  This turned out to be a
bug in MIT Kerberos, which is now avoided by using a different library
API call that doesn't have that bug.  This workaround has now been
removed, so the error reporting from WebLogin on password change will
now be more accurate.