Rewrite and expand the keyring functions in Perl WebAuth module
The keyring manipulation functions of the WebAuth Perl module have
been rewritten to be object-oriented, introducing new WebAuth::Keyring
and WebAuth::KeyringEntry objects.  Perl code using the WebAuth module
to manipulate keyrings will have to be modified, since several
functions were removed in favor of the new interface.  Methods to
remove a key from a keyring, get the timestamps and keys associated
with keyring entries, and choose the best key from a keyring have been
added.

There is new documentation for the keyring manipulation objects and
methods, and a new test suite that more completely exercises the
keyring interface.

Change-Id: I50b8d59b558a05fc31e57572b47e1458aa58cdcb
Reviewed-on: https://gerrit.stanford.edu/12
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>