Update the WebLogin flow diagram
If the user requests REMOTE_USER from the main login screen, we POST
back to the appropriate URL rather than do a redirect.  Adjust the
diagram accordingly.  Show that the confirmation screen can be forced
after single sign-on or REMOTE_USER authentication as well, and add the
suspicious login case to the list of reasons why we may force the
confirmation screen.  Add the correct arrow after login for a successful
login that doesn't require multifactor.

Add an SVG version of the document.

Change-Id: Iff9726e0f6fb7b4fcc13e4302aa50f8d51e0b893
Reviewed-on: https://gerrit.stanford.edu/22
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>