Communicate initial authentication factor information
Set the initial authentication factors in mod_webkdc (currently
hard-coded to password) and store that information in the webkdc-proxy
and generated id tokens.  Read that information in mod_webauth and
store it in the app token, along with session authentication factors
and LoA (if present, which they won't be yet).  Set environment variables
containing that information if it's available.

Change-Id: I7ae5cd61e917923e64d8b31092f676ebb1220d1b
Reviewed-on: https://gerrit.stanford.edu/32
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>