Begin refactoring of WebAuth library and token parsing
The long-term goal is to move high-level token parsing from the
Apache modules and separate Perl code into the libwebauth library.
This means adding an APR dependency so that we can use pools for
memory handling, which in turn requires a new WebAuth context and
initialization and teardown functions.  It also requires adding
APR and APR-Util dependencies in the libwebauth library.

Split the basic macros and required includes for uint32_t into
webauth/defines.h.  Update the build machinery accordingly.

Add webauth/basic.h to the installed headers for the interface to
create and destroy a WebAuth context.  Move webauth_error_message
into that header file along with the error codes.  Modify it to
take a context so that it can report a stored error message.  Update
callers for the right include and to pass in NULL for that argument.

Add the start of a new token parsing API that handles high-level
token parsing, and put the prototypes and token structures into
webauth/tokens.h.  Implement decoding app tokens and add a test for
that decoding.

Probe for APR and APR-Util separate from probing for Apache libraries
for the modules, and link libwebauth with both libraries.

Change-Id: I835ccf315182cd57c283084fa7bab11337503ca0
Reviewed-on: https://gerrit.stanford.edu/35
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>