Rename CRED_SERVER to CRED_SERVICE
Rename the symbolic name for the service attribute of a credential
token to service instead of server.  This holds the Kerberos
principal to which the credentials can authenticate, and service is
the correct Kerberos terminology for this.

Not all field and structure names have been updated, since those
structures will go away once module code is converted to use the new
token format.

Change-Id: I4a942deb73b70708b7c97bdec9cc39b2a504df09
Reviewed-on: https://gerrit.stanford.edu/45
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>