Add library support for encoding app tokens
Add support for encoding app tokens to the new high-level section of
the libwebauth library, and add a test case.  Validate the data when
doing the encoding.  Also fix some incorrect comments in the token
decoding files, close a memory leak in the token decoding test, and
change the signature of the token encoding functions to reflect that
they promise not to change the provided structs.

Change-Id: I300de0f86f2e33e1c15b96a8a39b83c130a6118c
Reviewed-on: https://gerrit.stanford.edu/53
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>