Rewrite mod_webauth to use the new token routines
Replace all the hand-crafted token encoding and decoding code in
mod_webauth with calls to the new token encoding and decoding
functions in libwebauth.  Remove the mod_webauth-specific token
structures in favor of using the generic ones (except for the
service token cache, which is still serving other purposes).

Ensure that the new factor and LoA information is present and
tracked across all the various token types.  Extract that
information from proxy tokens directly now that proxy tokens also
contain that information.

Change-Id: I2219f9474b27bf776d604e56291ce13470a05578
Reviewed-on: https://gerrit.stanford.edu/73
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>