Improve WebAuth library errors, refactor token tests
Change the library error message to have the generic message and then
the specific detail following in parentheses.  This seems to make the
most sense for how the error messages are being constructed.

Refactor the token encoding and decodingn tests while taking this
change into account to avoid more repeated code.

Change-Id: I72b84b840f40d2fb5dc0001e4d7d5d92d377376e
Reviewed-on: https://gerrit.stanford.edu/58
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>