Add factor and level of assurance to proxy tokens
It's convenient in the mod_webauth code, and faster due to fewer
round trips, for the proxy token to contain all the same information
as a webkdc-style id token.  Add the factor and level of assurance
information to the proxy token encoding and decoding routines in the
WebAuth library and to the protocol specification.

Also explain more in the protocol specification about how proxy
tokens are used.

Change-Id: I4e948426d636a78e59754bd67c6c6277786bac86
Reviewed-on: https://gerrit.stanford.edu/61
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>