Add numbered X.509 authentication factors
As requested by Tom Jones.  Similar to OTP, there are a lot of ways
of handling X.509 factors, and a site may deploy several of them.
Add the possibility of numbered factors similar to what's used for
OTP.  Also add examples of numbered OTP and X.509 factors to the
protocol specification.

Change-Id: I1b1bbc935685d2da022862290ab003858ee4f348
Reviewed-on: https://gerrit.stanford.edu/62
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>