Add support for app tokens containing only a session key
We reuse app tokens to convey application state through a WebKDC
request to handle returning to a different host in a pool than the
originating host.  Adjust token encoding and decoding to support
app tokens that contain only a session key.  Update the Perl library
documentation to include the session_key method, which wasn't
previously documented.

Change-Id: Ie2dba3ef600daa07043596511ee5a16738c50b93
Reviewed-on: https://gerrit.stanford.edu/66
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>