Provide an APR-aware version of webauth_context_init
For applications that have already initialized APR, provide a version
of webauth_context_init that requires the pool argument and does not
call apr_initialize.  This does not have to be (and should not be)
paired with a call to webauth_context_free; instead, the calling
application can just clear or destroy the parent pool and forget about
the WebAuth context.

Change-Id: Ifb1806abd6d0cfaaa2b879588ea84fe020ece719
Reviewed-on: https://gerrit.stanford.edu/67
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>