Add function to generate a WebAuth keyring from one key
Several places in the mod_webauth code we want to encode or decode
a token from a specific key.  Rather than add additional entry points
for the token encoding and decoding functions that take keys, provide
a function to convert a key into a keyring.

This function uses pool-allocated memory, which means one has to be
very careful with the resulting keyring and not call existing keyring
functions on it that may resize it, since that will fail horribly.
Eventually, all key management functions will be converted to use APR.

Change-Id: I7ef9eb58edffb4033948a884fa66236538b8b11f
Reviewed-on: https://gerrit.stanford.edu/68
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>