Change the WebAuth symbol version
The non-backward-compatible change in webauth_error_message means
that calls to WEBAUTH_2.0 version of that symbol will result in
segfaults and other errors.  Bump the global symbol version of all
symbols so that, if two libwebauth libraries are loaded at the
same time, there won't be a risk of getting the wrong symbol and
segfaulting.

Change-Id: I702c927d07e4234354392721d3004f611f53223a
Reviewed-on: https://gerrit.stanford.edu/74
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>