Fix warnings and hangs in the Perl module test suite
Override the warning for redefining a sub in t/weblogin.t.  Be sure
to pass the empty hash reference to all creations of CGI objects
so that they don't block trying to read from standard input.  Pass
in the existing CGI query object to WebLogin via QUERY so that it
doesn't attempt to create a new CGI object that reads from standard
input.

Change-Id: Iea8e78f8bb06d449c5adc49b122015b4b665e3b5
Reviewed-on: https://gerrit.stanford.edu/80
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>