Add a WebLogin hook function to specify factors for REMOTE_USER
Define and call a new WebLogin hook function that specifies the
authentication factors for REMOTE_USER authentications.  Use the
initial factor and level of assurance information when forming the
WebKDC proxy token.  The session information is not currently used.

Set the initial authentication factors to u (unknown) for
REMOTE_USER authentication if the hook function is not defined.

Change-Id: I50d6684a582760ceb6b9a68c61ca12398156e590
Reviewed-on: https://gerrit.stanford.edu/83
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>