Add initial and session factors of u to mod_webkdc
For right now, set the session factors in proxy and id tokens to
u and set the initial factors for a webkdc-proxy token obtained
via the XML interface to u.  The former will require additional
development work to understand the difference between session and
initial credentials, and the latter will require later configuration
work, but this change allows completion of testing of mod_webauth's
directives.

Change-Id: If15cae76f2fca3bb8fad97130febeb60ee0a6aa3
Reviewed-on: https://gerrit.stanford.edu/85
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>