Add multifactor page flow to WebLogin
Update WebLogin to understand how to handle a request that requires
multifactor.  Push the user to a new template page that will request a
one-time password and, if configured, allow the user to request a
one-time password be sent by some external mechanism.  Pass the OTP back
to the WebKDC code for validation and then either continue on to
confirmation or return to the multifactor page on failure.

Change-Id: I34a60b8b126af26e0629f7009e9e321216156c87
Reviewed-on: https://gerrit.stanford.edu/90
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>