Add support for querying a metadata service for user metadata
Add the beginnings of a set of WebKDC support functions to the
libwebauth library.  The first function queries a metadata service
for metadata about a particular user (particularly multifactor
information for the moment).  Currently, the only supported protocol
is remctl.

Add a generic APR string buffer type to libwebauth for internal use.

Add remctl to the build system and add test suite machinery for
starting a remctld server for running test suites.  Add a test suite
for the new metadata query routines.

Change-Id: Idf7d21650c5875723c1cba9ef8d02d7675073203
Reviewed-on: https://gerrit.stanford.edu/99
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>