Further WebKDC protocol updates for multifactor
Remove the type attribute of <subjectCredential> and instead allow
both <loginToken> and <proxyToken> in the same request, since a
multifactor authentication will require sending both.

Change the definition of <configuredFactor> to include just those
factors that the user has configured, satisfy the WAS requirements,
and has not already established.  Allow it to be multivalued.

Add a source attribute to <proxyToken> to specify where the token
came from, which will be used to establish session authentication
factors.

Include the type in <proxyToken> attributes, matching the current
implementation.

Change-Id: I1894e7e14dbaaef505a77e36b4357e9eb0aa3f5d
Reviewed-on: https://gerrit.stanford.edu/108
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>