Update WebLogin cookie handling for empty cookies
When logging out, we were setting the cookie value to '' while setting
the expiration time to the past.  Setting the value to '' isn't valid
for the HTTP protocol.  Instead, set it to a value, then check and skip
any cookies with that value.

Also check for any proxies we are turning into cookies, and do the same
thing if any of those proxies have an empty string for the value.  Empty
string is used to clear an invalid proxy token, so we want to expire
those the same way.

Change-Id: I85c5c54426389f377c84fccd399887c9fa3cd043
Reviewed-on: https://gerrit.stanford.edu/110
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>