Add mod_webkdc support for user metadata and multifactor
Add support for querying for user metadata if so configured.  Add
support for fulfilling multifactor requirements and tracking the
initial factors through multiple webkdc-proxy tokens into the tokens
generated for the WAS.  Add support for rejecting logins with
multifactor required or multifactor unavailable based on the user's
configuration and the requests from the WAS.

This provides the core of basic multifactor support except that it
provides no way of getting webkdc-proxy tokens with the appropriate
factors.  That will be coming in a subsequent change.

Most of the login code is moved into the libwebauth library, but there
are a lot of remnants still left behind in mod_webkdc that will need
to move later, and therefore a lot of duplicate code.

Change-Id: I4d10a121472d93caa8e00e5b755949f862e71f82
Reviewed-on: https://gerrit.stanford.edu/118
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>