Obtain Kerberos tickets before making remctl calls
When making user metadata calls with the remctl protocol, obtain
Kerberos tickets with the configured keytab first.  Pass the
configured keytab for the WebKDC module from the module into the
user metadata configuration.

This implementation currently uses environment variables, which both
has a memory leak of a small amount of memory on each pass through
the user metadata handler and changes Apache global process state
in undesireable ways.  Fixing this will require changes to the remctl
library.

Also introduce a webauth_krb5_get_cache function to retrieve the name
of the Kerberos ticket cache used for a WebAuth Kerberos context.

Change-Id: Ic7189e85b419bbb533fef22bc5df222e23c405e9
Reviewed-on: https://gerrit.stanford.edu/142
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>