Updates for working multifactor login
We weren't reloading various initialization tasks on several runmodes
that required them.  We weren't loading the proxy cookies on the
multifactor screen.  When constructing the WebKDC request, we weren't
setting the OTP value properly and using an incorrect method of going
through the set of proxy cookies.  We weren't processing the username
correctly through the multifactor flow.  We also weren't retrieving
the username from an existing proxy token if the user didn't have to
enter a password.  And we weren't setting the username and OTP when
doing the WebKDC request.

Change-Id: I90849aeb1236d26806ba4cc3d2fff3dcf6b8bc8e
Reviewed-on: https://gerrit.stanford.edu/149
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>