When merging webkdc-proxy tokens, use the oldest creation
We decide whether we can lift initial factors into session factors
based on the creation time of the webkdc-proxy token, so when we
merge tokens together, we need to use the oldest creation time of
the tokens of interest.  Otherwise, we keep renewing the token time
and it keeps being eligible for contributing to session factors.

Change-Id: Id66cd0804f6cf1b49f5ec1f313704ffc3434105b
Reviewed-on: https://gerrit.stanford.edu/163
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>