Support random multifactor in parsing code
Set a top-level flag in the factors struct if random multifactor
is present as a factor.  In the subset code, allow multifactor to
satisfy random multifactor when computing whether one set of factors
is a subset of another.  When parsing a random multifactor factor
and adding it to an existing set, ignore it if multifactor is already
present.

Change-Id: Idea5a77d365a35a39a383ed218355cfc801aada2
Reviewed-on: https://gerrit.stanford.edu/167
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>