Fixed display of insufficient multifactor error page
The insufficient multifactor error page shows the requested multifactor
types from the requesting site.  These were stored in an array, which we
were attempting to print, causing an error.  Make the default template
able to iterate through the array and display each member.  This way,
site-specific templates can override the o* level names with more
descriptive text in the template.

Change-Id: I8771816f5cbbd047fe315e499598fff730a9ae9c
Reviewed-on: https://gerrit.stanford.edu/170
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>