Fixed error check on if user had multifactor configured
In deciding whether the user had multifactor unconfigured or if the user
had multifactor configured but insufficient, we were checking to see if
factor_configured was defined.  It will be defined no matter what.
Change the WebResponse object to hold the entire list rather then just
the last value, then search that list to see if 'm' is an option.  Also
change the factor_type sent to the multifactor page to search for any
'o*' type, to follow the WebResponse change.

Change-Id: If5704e02456ba760ccfec9091b5492f6eec33926
Reviewed-on: https://gerrit.stanford.edu/171
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>