Update to rra-c-util 4.2 and C TAP Harness 1.10
Update to rra-c-util 4.2:

* Improve the xstrndup utility function.
* Simplify the test suite calls for Kerberos and remctl tests.

Update to C TAP Harness 1.10:

* Add bstrndup function to the C TAP library.

Eliminate a bunch of complexity in the Kerberos and remctl tests
to take advantage of the new TAP framework.  Remove the chdirs in
all the C code in favor of generating the remctld configuration
with Autoconf and adding a chdir to the backend command it runs.

Change-Id: Idfa75029eebef378edc177ac3528bfe1cb7adb85
Reviewed-on: https://gerrit.stanford.edu/250
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>