Add new WebKdcUserInfoIgnoreFail option
Add new mod_webkdc configuration option WebKdcUserInfoIgnoreFail,
which if set tells mod_webkdc to not fail the login if a user
information service is configured but cannot be queried (either due to
timeout or due to some other error).  By default, all logins will be
rejected if a user information service is configured but returns an
error.  If this option is set, the login can proceed, but only a
password factor will be available and no level of assurance can be
set.  Be aware that setting this option may allow bypassing a
multifactor requirement expressed by the user information service
rather than the destination site.

Change-Id: I3f61fb1a0d3f5a5226c2f4b4a63ac83d6468a2e5
Reviewed-on: https://gerrit.stanford.edu/255
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>