Document Apache/Tomcat interaction on URL parsing
Document in the mod_webauth manual a problematic interaction of URL
parsing between Apache and Tomcat that affects any Apache
authentication mechanism used to protect URLs that are proxied to
Tomcat.  Apache configuration to restrict access to proxied URLs needs
to allow for URI path parameters at the end of path segments.

Change-Id: I93fdd3a79f3369b3d4048394f55ad8fa434c0feb
Reviewed-on: https://gerrit.stanford.edu/266
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>