Initial ugly port to Apache 2.4
The minimum changes required for the modules to compile and work with
Apache 2.4 and pass basic testing.  This breaks compilation on Apache
2.2, which will be restored in a subsequent commit with additional
Autoconf probing.

This duplicates a bunch of code in mod_authldap and removes old code
that will no longer compile with #if 0.  This too will be cleaned up
in a subsequent release.

Change-Id: Ied003433a40c705d3fc9b78c25859bd9b243fb32
Reviewed-on: https://gerrit.stanford.edu/292
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>