Finish port to Apache 2.4
Add Autoconf probes to restore previous behavior when built with
Apache 2.2 or earlier.  Add a portable/apache.h header to encapsulate
the various portability fixes and the inclusion of a basic set of
Apache headers.  Adjust all Apache module code to use that header.

Move modules/mod-config.h to config-mod.h at the top level.

Add a NEWS entry for the Apache 2.4 port and warn that the legacy
StanfordAuth support is not available in mod_webauthldap when built
with Apache 2.4.

Change-Id: I8dbc7d70d06759e51da363e7568e9b779d7b6c86
Reviewed-on: https://gerrit.stanford.edu/293
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>