Fix merging of mod_webkdc Apache directives in multiple merges
Fix merging of mod_webkdc Apache directives in some corner cases where
the directive has a default value or is explicitly set to off.

Previously, the _set flag was not set on the newly generated config
after a merge of a set value, which could cause the winning value to be
discarded in a later merge.

Change-Id: If4244b1340b28760f6fee45e9a9dd13d619328cd
Reviewed-on: https://gerrit.stanford.edu/295
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>