Log authorization denied at debug level for Apache 2.4
In Apache 2.4, we might have a bunch of separate directives or
be called in complex ways, so it doesn't make sense to log the
authorization denied state at a high priority level.  Previously,
that made sense because mod_webauthldap handled the complete group
list at once, but that's no longer the case.  Reduce the level to
debug.

Change-Id: I74fa215d50bd4a0c1a47f6fd90dcaf889bf8e34a
Reviewed-on: https://gerrit.stanford.edu/303
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>