Rewrite the Perl API to create a WebAuth object for the context
The WebAuth Perl module API now requires creating a WebAuth object
first and passing that object as the first argument to most other
functions.  This is the first step in making the API more
object-oriented.  All users will need code changes to work with the
new API.  WebAuth::Keyring and WebAuth::Krb5 have not yet been
converted, but will be in a subsequent release.  This means that the
WebKDC and WebLogin Perl modules in this release require the WebAuth
module from this release and vice versa, so be careful of partial
upgrades.

Change-Id: I30ca20da55ed250d18057d65f762071cb7d0b4ae
Reviewed-on: https://gerrit.stanford.edu/305
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>