Add WebAuth::Token::App and decoding app tokens
Add a new framework for token decoding that uses the new libwebauth
API and then maps the resulting token to a Perl hash.  Create a
WebAuth::Token::App class to wrap app tokens and add decoding and a
test for app tokens.

This also adds a test framework that starts duplicating some of the
TAP library functions for C and shell in a Perl module framework.

Change-Id: I12df89cc9b5dbcdb69edbfbab9ca90931387f641
Reviewed-on: https://gerrit.stanford.edu/309
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>