New test token configuration based on generate-tokens
As the first step of shifting the Perl test suite to be more
data-driven, create a new configuration file that lists all of our
test tokens.  This is like the token data in generate-tokens, except
using the new WebAuth::Token::* class names and with attribute names
that match the struct members for the webauth_token_* structs and
therefore our accessor functions.

Divide the test tokens into three groups: ones that are good, ones
that are syntactically valid but will produce errors on decoding,
and ones that are syntactically invalid.

Change-Id: I77030cc52f45cbcf89e0bdaec1ca5dbe8272dc84
Reviewed-on: https://gerrit.stanford.edu/314
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>