Add new webauth-make-tokens script based on generate-tokens
Add new tools/webauth-make-tokens script to generate WebAuth tokens
given a configuration file and keyring.  This is not installed by
default and is normally only used to generate test data, but it may be
useful in some other cases of manual token generation.

This doesn't replace generate-tokens yet, since not all tokens are
supported by the new library and there isn't yet support for generating
the various invalid tokens.

Change-Id: Ie0a2b505d465b2101ff10b341cf767625beda7fa
Reviewed-on: https://gerrit.stanford.edu/317
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>